refactor: Change all SpotifyAppRemote methods to static
Seems that on iOS we will have to instruct the developer to create a
custom AppDelegate so we can catch the token provided by Spotify when
the app is called back via the redirect URI, so the members being static
help to avoid losing state between the app initialization and the actual
page where the plugin is gonna be used.

Also it makes no sense having two or more instances of SpotifyAppRemote,
so static is the way to go.